requirements.txt
Django>=5.0
Pillow>=10.0
opencv-python>=4.8
pytesseract>=0.3.10
pandas>=2.0
openpyxl>=3.0
google-generativeai>=0.3.0
python-dotenv>=1.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
